Move courtnav mapper to mappers package
